EH-939: Moved ShareType and TutkinnonOsaType to ShareTypes.ts
