+++ shared/models/AiemminHankittuAmmatillinenTutkinnonOsa.ts
@@ -11,7 +11,7 @@ import { KoodistoVastaus } from "models/KoodistoVastaus"
 import { AiemminHankitutTutkinnonOsatViews } from "./helpers/AiemminHankitutTutkinnonOsatViews"
 import { Organisaatio } from "./Organisaatio"
 import { EnrichOrganisaatioOid } from "./EnrichOrganisaatioOid"
-import { TutkinnonOsaType } from "./helpers/TutkinnonOsa"
+import { TutkinnonOsaType } from "./helpers/ShareTypes"
 
 const Model = types.model({
   id: types.optional(types.number, 0),
