+++ shared/models/AiemminHankittuYhteinenTutkinnonOsa.ts
@@ -5,7 +5,7 @@ import { AiemminHankitunYTOOsaAlue } from "./AiemminHankitunYTOOsaAlue"
 import { EPerusteetVastaus } from "models/EPerusteetVastaus"
 import { TutkinnonOsaViite } from "models/TutkinnonOsaViite"
 import { KoodistoVastaus } from "models/KoodistoVastaus"
-import { TutkinnonOsaType } from "./helpers/TutkinnonOsa"
+import { TutkinnonOsaType } from "./helpers/ShareTypes"
 
 export const AiemminHankittuYhteinenTutkinnonOsa = types
   .model("AiemminHankittuYhteinenTutkinnonOsa", {
