+++ shared/models/HankittavaAmmatillinenTutkinnonOsa.ts
@@ -9,7 +9,7 @@ import { TutkinnonOsaViite } from "models/TutkinnonOsaViite"
 import { EnrichTutkinnonOsa } from "models/EnrichTutkinnonOsa"
 import { EnrichOrganisaatioOid } from "./EnrichOrganisaatioOid"
 import { Organisaatio } from "./Organisaatio"
-import { TutkinnonOsaType } from "./helpers/TutkinnonOsa"
+import { TutkinnonOsaType } from "./helpers/ShareTypes"
 
 export const Model = types.model({
   id: types.optional(types.number, 0),
