+++ shared/models/helpers/ShareTypes.ts
@@ -1,0 +1,15 @@
+export enum ShareType {
+  "osaamisenosoittaminen" = "osaamisenosoittaminen",
+  "osaamisenhankkimistapa" = "osaamisenhankkimistapa"
+}
+
+export enum TutkinnonOsaType {
+  HankittavaAmmatillinen = "HankittavaAmmatillinenTutkinnonOsa",
+  HankittavaPaikallinen = "HankittavaPaikallinenTutkinnonOsa",
+  HankittavaYhteinen = "HankittavaYhteinenTutkinnonOsa",
+  HankittavanYhteisenTutkinnonOsanOsaAlue = "HankittavaYTOOsaAlue",
+  AiemminHankittuAmmatillinen = "AiemminHankittuAmmatillinenTutkinnonOsa",
+  AiemminHankittuPaikallinen = "AiemminHankittuPaikallinenTutkinnonOsa",
+  AiemminHankittuYhteinen = "AiemminHankittuYhteinenTutkinnonOsa",
+  AiemminHankitunYhteisenTutkinnonOsanOsaAlue = "AiemminHankittuYTOOsaAlue"
+}
